Added Google Backend Auth

+++ Project-Tech/Backend/unlockit/src/main/java/io/unlockit/controller/AdvertiseController.java
@@ -1,11 +1,11 @@
 package io.unlockit.controller;
 
+import io.unlockit.google.GoogleUtils;
 import io.unlockit.model.mongodb.Advertise;
 import io.unlockit.service.AdvertiseService;
 import io.unlockit.utils.FrontendEndpoint;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,44 +20,68 @@ public class AdvertiseController {
     private AdvertiseService advertiseService;
 
     @PostMapping("/register")
-    public ResponseEntity<Advertise> registerAdvertise(@RequestBody @Valid Advertise advertise) {
-        Advertise newAdvertise = advertiseService.saveAdvertise(advertise);
-        return ResponseEntity.ok().body(newAdvertise);
+    public ResponseEntity<Advertise> registerAdvertise(@RequestBody @Valid Advertise advertise, @RequestHeader("Authorization") String authorizationHeader) {
+        if (GoogleUtils.isRequestAuthorized(authorizationHeader)) {
+            Advertise newAdvertise = advertiseService.saveAdvertise(advertise);
+            return ResponseEntity.ok().body(newAdvertise);
+        }
+        return ResponseEntity.badRequest().body(null);
     }
 
     @GetMapping("/all")
-    public ResponseEntity<List<Advertise>> getAllAdvertises() {
-        List<Advertise> advertises = advertiseService.getAllAdvertises();
-        return ResponseEntity.ok().body(advertises);
+    public ResponseEntity<List<Advertise>> getAllAdvertises(@RequestHeader("Authorization") String authorizationHeader) {
+        if (GoogleUtils.isRequestAuthorized(authorizationHeader)) {
+            List<Advertise> advertises = advertiseService.getAllAdvertises();
+            return ResponseEntity.ok().body(advertises);
+        }
+        return ResponseEntity.badRequest().body(null);
     }
 
     @GetMapping("/user/get/all/{id}")
-    public ResponseEntity<List<Advertise>> getAllAdvertisesByUserId(@PathVariable("id") String userId) {
-        List<Advertise> advertises = advertiseService.getAllAdvertisesByUserId(userId);
-        return ResponseEntity.ok().body(advertises);
+    public ResponseEntity<List<Advertise>> getAllAdvertisesByUserId(@PathVariable("id") String userId, @RequestHeader("Authorization") String authorizationHeader) {
+        if (GoogleUtils.isRequestAuthorized(authorizationHeader)) {
+            List<Advertise> advertises = advertiseService.getAllAdvertisesByUserId(userId);
+            return ResponseEntity.ok().body(advertises);
+        }
+        return ResponseEntity.badRequest().body(null);
     }
 
     @GetMapping("/location/get/all/{id}")
-    public ResponseEntity<List<Advertise>> getAllAdvertisesByLocation(@PathVariable("id") String location) {
-        List<Advertise> advertises = advertiseService.getAllAdvertisesByLocation(location);
-        return ResponseEntity.ok().body(advertises);
+    public ResponseEntity<List<Advertise>> getAllAdvertisesByLocation(@PathVariable("id") String location, @RequestHeader("Authorization") String authorizationHeader) {
+        if (GoogleUtils.isRequestAuthorized(authorizationHeader)) {
+            List<Advertise> advertises = advertiseService.getAllAdvertisesByLocation(location);
+            return ResponseEntity.ok().body(advertises);
+        }
+        return ResponseEntity.badRequest().body(null);
     }
 
     @GetMapping("/get/{id}")
-    public ResponseEntity<Advertise> getAdvertiseById(@PathVariable("id") String advertiseId) {
-        Advertise advertise = advertiseService.getAdvertiseById(advertiseId);
-        return ResponseEntity.ok().body(advertise);
+    public ResponseEntity<Advertise> getAdvertiseById(@PathVariable("id") String advertiseId, @RequestHeader("Authorization") String authorizationHeader) {
+        if (GoogleUtils.isRequestAuthorized(authorizationHeader)) {
+            Advertise advertise = advertiseService.getAdvertiseById(advertiseId);
+            return ResponseEntity.ok().body(advertise);
+        }
+        return ResponseEntity.badRequest().body(null);
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<Advertise> updateAdvertise(@RequestBody @Valid Advertise advertise, @PathVariable("id") String advertiseId) {
-        Advertise updatedAdvertise = advertiseService.updateAdvertise(advertise, advertiseId);
-        return ResponseEntity.ok(updatedAdvertise);
+    public ResponseEntity<Advertise> updateAdvertise(@RequestBody @Valid Advertise advertise, @PathVariable("id") String advertiseId, @RequestHeader("Authorization") String authorizationHeader) {
+        if (GoogleUtils.isRequestAuthorized(authorizationHeader)) {
+            Advertise updatedAdvertise = advertiseService.updateAdvertise(advertise, advertiseId);
+            return ResponseEntity.ok(updatedAdvertise);
+        }
+        return ResponseEntity.badRequest().body(null);
     }
 
     @DeleteMapping("/delete/{id}")
-    public HttpStatus deleteAdvertise(@PathVariable("id") String advertiseId) {
-        advertiseService.deleteAdvertise(advertiseId);
-        return HttpStatus.OK;
+    public ResponseEntity<Void> deleteAdvertise(@PathVariable("id") String advertiseId, @RequestHeader("Authorization") String authorizationHeader) {
+        if (GoogleUtils.isRequestAuthorized(authorizationHeader)) {
+            advertiseService.deleteAdvertise(advertiseId);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.badRequest().build();
     }
+
+
+
 }
